Separate UI and business logic
Move all code except for business logic into index.js.
